first way of mechanical energy but think there's a problem, should not be a shear term

+++ sim1dplotting/sim1dplotting/utils/TimeUtilities.py
@@ -5,7 +5,7 @@ class TimeUtility:
     def __init__(self, configuration_time = None, 
                  configuration_fig = None):
         
-        self.dt           = int(configuration_time['dt'])
+        self.dt           = float(configuration_time['dt'])
         self.outputdir     = str(configuration_fig['outputdir'])
